+++ src/itemActionTitle/index.tsx
@@ -1,0 +1,2 @@
+export { ItemActionTitle, ItemActionTitleProps } from './ItemActionTitle'
+export { ItemActionTitle as default } from './ItemActionTitle'