+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- **[NEW]** Add `ItemActionTitle`
+- **[UPDATE]** Provide `title` attribute to `A11yProps`
 - [...]
 
 # v41.9.4 (30/11/2020)